stopped using gamestatecontext
The resetGame function was initially part of the GameStateContext, which is a global state management solution. However, the resetGame function was not actually resetting the game state when called from different components, such as the NavigationButtons component.

To fix this, we moved the resetGame function from the GameStateContext to the individual components where it's needed (i.e., the Home and HistoryPage components). We then passed the resetGame function as a prop to the NavigationButtons component. This ensures that the resetGame function has access to the local state of the component where it's defined and can effectively reset the game state.

This change also makes the data flow in the app more explicit and easier to understand. While context is useful for sharing data that can be considered "global" for a tree of React components, it's not necessary in this case because the resetGame function is only needed in specific components. By passing the resetGame function as a prop, we can clearly see where it's coming from and where it's being used.