Revert _app.js to Previous Version Without next/head
This commit reverts the changes made to _app.js, specifically removing the next/head import and its usage. The previous version of the file did not include next/head, which was introduced to add a favicon link to the application's head. This revert is being done to troubleshoot a 'MODULE_NOT_FOUND' error encountered in production, specifically related to 'next/dist/server/future/route-modules/pages/vendored/contexts