Temporarily comment out next/head in layout.js
This commit temporarily comments out the usage of next/head in layout.js. This is a troubleshooting step to determine if next/head is causing a module not found error during server-side rendering on the results page. The usage of next/head will be restored once the issue is resolved.