Implement Anonymous User Flow with Leaderboard Name
This commit introduces the ability for users to play the game anonymously by choosing a leaderboard name. The leaderboard name is used for Firebase anonymous sign-in and is stored in the Firestore database. At the end of the game, users are prompted to create an account to save or share their score, which upgrades their anonymous account to a permanent account using Firebase's linkWithCredential function.