Handle Firestore updates for anonymous users
This commit introduces changes to handle Firestore document updates for anonymous users. The GameOver component now receives the user object as a prop, which is used to determine whether the user is anonymous or not. Based on this, the correct Firestore document (either in the 'users' or 'guests' collection) is updated. This resolves the issue of "No document to update" error when an anonymous user finishes a game.