Implement Streaming and Suspense for Leaderboard Page
This commit introduces the use of React Suspense and streaming to improve the performance of the leaderboard page. Initially, only the first page of data for the 'lifetime' filter is fetched and rendered. Additional data is then fetched in the background and stored in state, ready to be displayed when the user interacts with the page. React Suspense is used to show a loading state while this additional data is being fetched. This approach improves the perceived performance of the page and provides a smoother user experience.