Add Assistant Settings to Options Menu
This commit adds a new option for Assistant Settings in the Options Menu. When this option is clicked, the user is navigated to the Assistant Settings page. If the user is not logged in, the authentication modal is shown. This allows users to easily access and modify their assistant settings directly from the Options Menu.