Remove manual setting of isAuthLoading state
Removed instances of setIsAuthLoading(false) from the resetGame and handleUserUpdate functions in pages/index.js. The isAuthLoading state is managed by the useAuth context and should automatically update when the authentication status changes. Manually setting this state is not necessary and was causing a ReferenceError because setIsAuthLoading was not defined.