Remove unused handleUpgradeAccount function
This commit removes the handleUpgradeAccount function from the Auth component. After a thorough review of the codebase, it was determined that this function was not being used anywhere in the application. This function was a remnant from an older version of the application where guest accounts could be upgraded to regular accounts. However, this functionality is no longer needed as the application now handles all users in a single collection.