Refactor import paths and add instructions for adding new functions
This commit updates the import paths in index.js to correctly reference the individual function files in the same directory. Previously, the paths were incorrectly prefixed with ./functions/, which caused the Firebase Emulator Suite to fail to load the functions.

Additionally, this commit adds detailed instructions in index.js on how to add new functions. These instructions guide the developer through the process of creating a new function file, writing and exporting the function, importing and re-exporting the function in index.js, testing the function locally with the Firebase Emulator Suite, and finally deploying the function to Firebase.