Add button click sound effect to OptionsMenu component
This commit introduces a button click sound effect to the OptionsMenu component. The sound effect is played when the options menu button and any of the dropdown items are clicked. The 'use-sound' library is used to handle the sound effect. The sound file is stored in the public directory and is loaded from there when needed.