Add game start sound effect in Home component
This commit introduces a game start sound effect in the Home component. The sound effect is played when a game mode is selected, which signifies the start of the game. The 'use-sound' library is used to handle the sound effect. The sound file is stored in the public directory and is loaded from there when needed.