Add feedback subfield to customInstructions and display feedback to user
This commit introduces a new subfield feedback under the customInstructions object. This subfield is used to store feedback from the OpenAI API regarding the safety of custom instructions provided by the user.

The getServerSideProps function in assistantSettings/[userId].js has been updated to fetch the feedback subfield along with the rest of the user data from Firestore.

A new state variable feedback has been added to the AssistantSettingsPage component, and it is updated whenever the customInstructions.feedback prop changes.

Finally, a toast notification is displayed to the user whenever the feedback state variable changes. This provides the user with immediate feedback about the safety of their custom instructions.