Clear feedback field in Firestore after toast display
Added code to clear the customInstructions.feedback field in Firestore after the toast is displayed. This is done in the useEffect hook in pages/assistantSettings/[userId].js that triggers the toast. The feedback field is cleared in Firestore 4 seconds after the toast is displayed, which is the default duration for a 'blank' toast type.